clean then put
